requirements.txt
pandas >= 1.5.0
numpy >= 1.23.0

scikit-learn >= 1.2.0
matplotlib >= 3.6.0
umap-learn >= 0.5.2
InstructorEmbedding >= 0.1.0
PyYAML >= 6.0
faiss-cpu >= 1.7.2